SC-23 Add authorization check in deleteRoomController

+++ src/controllers/room/delete/deleteRoomController.ts
@@ -3,16 +3,23 @@ import { Inject, Service } from "typedi";
 import config from "../../../../config";
 import IDeleteRoomController from "../../IControllers/room/delete/IDeleteRoomController";
 import IDeleteRoomService from "../../../services/IServices/room/IDeleteRoomService";
+import { IAuthService } from "../../../services/IServices/auth/IAuthService";
 
 @Service()
 export default class DeleteRoomController implements IDeleteRoomController {
 
     constructor(
-        @Inject(config.services.deleteRoom.name) private service: IDeleteRoomService
+        @Inject(config.services.deleteRoom.name) private service: IDeleteRoomService,
+        @Inject(config.services.auth.name) private authService: IAuthService
     )
     {}
 
     public async deleteRoom(req: Request, res: Response, next: NextFunction) {
+        //@ts-ignore
+        let userRole = req.userRole;
+        if(!this.authService.validatePermission(userRole, ["CampusManager"])){
+            return res.status(401).send("Unauthorized");
+        }
         try {
             const roomId = String(req.params.roomName);
             const deletedOrError = await this.service.deleteRoom(roomId)
